[ContentFragment] Move content fragment component from extension to core (#493)
This moves models, servlets and their tests to the core components.
Newly required dependencies are added/adjusted.

Mass-format code and optimize imports.
