+++ bundles/core/src/test/java/com/adobe/cq/wcm/core/components/ContentFragmentModelsCommonsTest.java
@@ -1,0 +1,31 @@
+/*~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+ ~ Copyright 2017 Adobe Systems Incorporated
+ ~
+ ~ Licensed under the Apache License, Version 2.0 (the "License");
+ ~ you may not use this file except in compliance with the License.
+ ~ You may obtain a copy of the License at
+ ~
+ ~     http://www.apache.org/licenses/LICENSE-2.0
+ ~
+ ~ Unless required by applicable law or agreed to in writing, software
+ ~ distributed under the License is distributed on an "AS IS" BASIS,
+ ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ ~ See the License for the specific language governing permissions and
+ ~ limitations under the License.
+ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
+package com.adobe.cq.wcm.core.components;
+
+import org.junit.Test;
+
+import com.adobe.cq.wcm.core.components.testing.AbstractModelTest;
+
+public class ContentFragmentModelsCommonsTest extends AbstractModelTest {
+
+    @Test
+    public void testDefaultBehaviour() throws Exception {
+        testDefaultBehaviour(new String[] {
+                "com.adobe.cq.wcm.core.components.models.v1.contentfragment"
+        });
+    }
+
+}
