+++ bundles/core/src/test/resources/contentfragment/test-content.json
@@ -1,0 +1,121 @@
+{
+    "contentfragments": {
+        "jcr:primaryType": "cq:Page",
+        "jcr:content"    : {
+            "jcr:primaryType"   : "cq:PageContent",
+            "jcr:title"         : "Content Fragments Test",
+            "sling:resourceType": "core/wcm/components/page/v2/page",
+            "root"              : {
+                "jcr:primaryType"   : "nt:unstructured",
+                "sling:resourceType": "wcm/foundation/components/responsivegrid",
+                "responsivegrid"    : {
+                    "jcr:primaryType"                      : "nt:unstructured",
+                    "sling:resourceType"                   : "wcm/foundation/components/responsivegrid",
+                    "text-only-no-path"                    : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment"
+                    },
+                    "text-only-non-existing-path"          : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/non-existing"
+                    },
+                    "text-only-invalid-path"               : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments"
+                    },
+                    "text-only"                            : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/text-only"
+                    },
+                    "text-only-variation"                  : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/text-only",
+                        "variationName"     : "teaser"
+                    },
+                    "text-only-non-existing-variation"     : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/text-only",
+                        "variationName"     : "non-existing"
+                    },
+                    "text-only-single-element"             : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/text-only",
+                        "elementNames"      : ["second"]
+                    },
+                    "text-only-multiple-elements"          : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/text-only",
+                        "elementNames"      : ["second", "non-existing", "main"]
+                    },
+                    "structured-no-path"                   : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment"
+                    },
+                    "structured-non-existing-path"         : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/non-existing"
+                    },
+                    "structured-invalid-path"              : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments"
+                    },
+                    "structured"                           : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/structured"
+                    },
+                    "structured-display-mode-single"                           : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/structured",
+                        "displayMode"       : "singleText"
+                    },
+                    "structured-variation"                 : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/structured",
+                        "variationName"     : "teaser"
+                    },
+                    "structured-non-existing-variation"    : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/structured",
+                        "variationName"     : "non-existing"
+                    },
+                    "structured-nested-model"              : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/structured-nested-model"
+                    },
+                    "structured-single-element"            : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/structured",
+                        "elementNames"      : ["second"]
+                    },
+                    "structured-single-element-main"       : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/structured",
+                        "elementNames"      : ["main"]
+                    },
+                    "structured-multiple-elements"         : {
+                        "jcr:primaryType"   : "nt:unstructured",
+                        "sling:resourceType": "core/wcm/components/contentfragment/v1/contentfragment",
+                        "fragmentPath"      : "/content/dam/contentfragments/structured",
+                        "elementNames"      : ["second", "non-existing", "main"]
+                    }
+                }
+            }
+        }
+    }
+}
