--- extension/contentfragment/bundle/src/test/resources/contentfragment/test-content-datasources.json
@@ -1,117 +1,0 @@
-{
-    "elements-no-config"                           : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/elements"
-    },
-    "elements-fragment-path-non-existing"          : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/elements",
-        "fragmentPath"      : "/content/dam/contentfragments/non-existing"
-    },
-    "elements-component-path-non-existing"         : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/elements",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/non-existing"
-    },
-    "elements-fragment-path-invalid"               : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/elements",
-        "fragmentPath"      : "/content/dam/contentfragments"
-    },
-    "elements-component-path-invalid"              : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/elements",
-        "componentPath"     : "/content/contentfragments"
-    },
-    "elements-component-path-invalid-fragment-path": {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/elements",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/text-only-invalid-path"
-    },
-    "elements-fragment-path-text-only"             : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/elements",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/text-only"
-    },
-    "elements-component-path-text-only"            : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/elements",
-        "fragmentPath"      : "/content/dam/contentfragments/text-only"
-    },
-    "elements-fragment-path-structured"            : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/elements",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/structured"
-    },
-    "elements-component-path-structured"           : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/elements",
-        "fragmentPath"      : "/content/dam/contentfragments/structured"
-    },
-    "elements-component-path-structured-display-mode-single"           : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/elements",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/structured-display-mode-single"
-    },
-    "elements-fragment-path-override"              : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/elements",
-        "fragmentPath"      : "/content/dam/contentfragments/structured",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/non-existing"
-    },
-    "variations-no-config"                           : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/variations"
-    },
-    "variations-fragment-path-non-existing"          : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/variations",
-        "fragmentPath"      : "/content/dam/contentfragments/non-existing"
-    },
-    "variations-component-path-non-existing"         : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/variations",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/non-existing"
-    },
-    "variations-fragment-path-invalid"               : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/variations",
-        "fragmentPath"      : "/content/dam/contentfragments"
-    },
-    "variations-component-path-invalid"              : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/variations",
-        "componentPath"     : "/content/contentfragments"
-    },
-    "variations-component-path-invalid-fragment-path": {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/variations",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/text-only-invalid-path"
-    },
-    "variations-fragment-path-text-only"             : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/variations",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/text-only"
-    },
-    "variations-component-path-text-only"            : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/variations",
-        "fragmentPath"      : "/content/dam/contentfragments/text-only"
-    },
-    "variations-fragment-path-structured"            : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/variations",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/structured"
-    },
-    "variations-component-path-structured"           : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/variations",
-        "fragmentPath"      : "/content/dam/contentfragments/structured"
-    },
-    "variations-fragment-path-override"              : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/datasource/variations",
-        "fragmentPath"      : "/content/dam/contentfragments/structured",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/non-existing"
-    }
-}
