--- extension/contentfragment/bundle/src/test/resources/contentfragment/test-content-renderconditions.json
@@ -1,14 +1,0 @@
-{
-    "display-mode-single-text"            : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/renderconditions/elementnames",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/structured-display-mode-single",
-        "displayMode"       : "singleText"
-    },
-    "display-mode-multi"             : {
-        "jcr:primaryType"   : "nt:unstructured",
-        "sling:resourceType": "core/wcm/extension/components/contentfragment/v1/renderconditions/elementnames",
-        "componentPath"     : "/content/contentfragments/jcr:content/root/responsivegrid/structured-display-mode-single",
-        "displayMode"       : "multi"
-    }
-}
