--- extension/contentfragment/bundle/src/test/resources/findbugs-exclude.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  ~ Copyright 2017 Adobe Systems Incorporated
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~-->
-<FindBugsFilter>
-    <Match>
-        <Bug pattern="SE_BAD_FIELD" />
-    </Match>
-    <Match>
-        <Bug pattern="SE_NO_SERIALVERSIONID" />
-    </Match>
-    <Match>
-        <Bug pattern="NM_SAME_SIMPLE_NAME_AS_INTERFACE" />
-    </Match>
-    <Match>
-        <Bug pattern="NM_SAME_SIMPLE_NAME_AS_SUPERCLASS" />
-    </Match>
-</FindBugsFilter>
