--- extension/contentfragment/bundle/src/test/resources/test-content.json
@@ -1,3 +1,0 @@
-{
-    ":type": "core/wcm/components/title/v1/title"
-}
