+++ pom.xml
@@ -70,7 +70,6 @@
         <module>config</module>
         <module>all</module>
         <module>examples</module>
-        <module>extension/contentfragment/bundle</module>
         <module>extension/contentfragment/content</module>
         <module>extension</module>
     </modules>
